EdgeSolution/modules/Scorer/requirements.txt
azure-iot-device~=2.7.0
setuptools
pyodbc
# onnxmltools
# onnxruntime
# setuptools
# sqlalchemy